gql_http_link: extract multipart files to client
