+++ links/gql_http_link/lib/gql_http_link.dart
@@ -3,3 +3,5 @@ library gql_http_link;
 
 export "package:gql_http_link/src/exceptions.dart";
 export "package:gql_http_link/src/link.dart";
+export "package:gql_http_link/src/gql_client.dart";
+export "package:gql_http_link/src/http_gql_client.dart";
