+++ links/gql_http_link/lib/src/_utils.dart
@@ -1,10 +1,7 @@
-import "dart:convert";
-import "dart:typed_data";
-import "package:meta/meta.dart";
-
-import "package:http/http.dart";
 import "package:gql/ast.dart";
 import "package:gql_exec/gql_exec.dart" as gql;
+
+typedef MultiPartFileTypeCheck = bool Function(dynamic file);
 
 extension WithType on gql.Request {
   OperationType get type {
@@ -24,7 +21,7 @@ extension WithType on gql.Request {
   bool get isQuery => type == OperationType.query;
 }
 
-/// Recursively extract [MultipartFile]s and return them as a normalized map of [path] => [file]
+/// Recursively extract files and return them as a normalized map of [path] => [file]
 /// From the given request body
 ///
 /// ```dart
@@ -36,17 +33,19 @@ extension WithType on gql.Request {
 ///   "foo.0.bar": MultipartFile("blah.txt")
 /// }
 /// ```
-Map<String, MultipartFile> extractFlattenedFileMap(
-  dynamic body, {
-  Map<String, MultipartFile> currentMap,
+Map<String, dynamic> extractFlattenedFileMap(
+  dynamic body,
+  MultiPartFileTypeCheck typeCheck, {
+  Map<String, dynamic> currentMap,
   List<String> currentPath = const <String>[],
 }) {
-  currentMap ??= <String, MultipartFile>{};
+  currentMap ??= <String, dynamic>{};
   if (body is Map<String, dynamic>) {
     final Iterable<MapEntry<String, dynamic>> entries = body.entries;
     for (final MapEntry<String, dynamic> element in entries) {
       currentMap.addAll(extractFlattenedFileMap(
         element.value,
+        typeCheck,
         currentMap: currentMap,
         currentPath: List<String>.from(currentPath)..add(element.key),
       ));
@@ -57,6 +56,7 @@ Map<String, MultipartFile> extractFlattenedFileMap(
     for (int i = 0; i < body.length; i++) {
       currentMap.addAll(extractFlattenedFileMap(
         body[i],
+        typeCheck,
         currentMap: currentMap,
         currentPath: List<String>.from(currentPath)..add(i.toString()),
       ));
@@ -64,44 +64,12 @@ Map<String, MultipartFile> extractFlattenedFileMap(
     return currentMap;
   }
 
-  if (body is MultipartFile) {
+  if (typeCheck(body)) {
     return currentMap
-      ..addAll({
+      ..addAll(<String, dynamic>{
         currentPath.join("."): body,
       });
   }
 
   return currentMap;
-}
-
-extension AddAllFiles on MultipartRequest {
-  void addAllFiles(Map<String, MultipartFile> fileMap) {
-    final Map<String, List<String>> fileMapping = <String, List<String>>{};
-    final List<MultipartFile> fileList = <MultipartFile>[];
-
-    final List<MapEntry<String, MultipartFile>> fileMapEntries =
-        fileMap.entries.toList(growable: false);
-
-    for (int i = 0; i < fileMapEntries.length; i++) {
-      final MapEntry<String, MultipartFile> entry = fileMapEntries[i];
-      final String indexString = i.toString();
-      fileMapping.addAll(<String, List<String>>{
-        indexString: <String>[entry.key],
-      });
-      final MultipartFile f = entry.value;
-      fileList.add(MultipartFile(
-        indexString,
-        f.finalize(),
-        f.length,
-        contentType: f.contentType,
-        filename: f.filename,
-      ));
-    }
-
-    fields["map"] = json.encode(fileMapping);
-
-    files.addAll(fileList);
-  }
-
-  set body(String body) => fields["operations"] = body;
 }
